anthropic-webui/requirements.txt
absl-py
anthropic
fastapi
openai
pydantic
pytest
pytest-asyncio
uvicorn
